backend/requirements.txt
flask
flask-cors
whisper
yt-dlp
groq
python-dotenv
